refactor: Move MachineIdForHistory DTO to dto/history-dto folder and rename to HistoryForMachineId
